+++ R/update.R
@@ -13,11 +13,13 @@
 #'
 #' @inheritParams quarto_render
 #'
-#' @param extension The extension to update, either an archive or a GitHub
-#'   repository as described in the documentation
-#'   <https://quarto.org/docs/extensions/managing.html>.
+#' @param extension The extension to update, either by its name (i.e ` quarto update extension <gh-org>/<gh-repo>`), an archive (` quarto update extension <path-to-zip>`) or a url (`quarto update extension <url>`).
 #'
-#' @param no_prompt Do not prompt to confirm approval to download external extension.
+#' @param no_prompt Do not prompt to confirm approval to download external extension. Setting `no_prompt = FALSE` means [Extension Trust](#extension-trust) is accepted.
+#'
+#' @seealso [quarto_add_extension()], [quarto_remove_extension()], and [Quarto website](https://quarto.org/docs/extensions/managing.html).
+#'
+#' @return Returns invisibly `TRUE` if the extension was updated, `FALSE` otherwise.
 #'
 #' @examples
 #' \dontrun{
@@ -27,9 +29,6 @@
 #' # Update a template and set up a draft document from a ZIP archive
 #' quarto_update_extension("https://github.com/quarto-ext/fontawesome/archive/refs/heads/main.zip")
 #' }
-#'
-#' @importFrom rlang is_interactive
-#' @importFrom cli cli_abort
 #' @export
 quarto_update_extension <- function(
   extension = NULL,
@@ -44,16 +43,20 @@ quarto_update_extension <- function(
   # This will ask for approval or stop installation
   approval <- check_extension_approval(
     no_prompt,
-    "Quarto extensions",
+    "Quarto extension",
     "https://quarto.org/docs/extensions/managing.html"
   )
 
-  if (approval) {
-    args <- c(extension, "--no-prompt", if (quiet) cli_arg_quiet(), quarto_args)
-    quarto_update(args, quarto_bin = quarto_bin, echo = TRUE)
+  if (!approval) {
+    return(invisible(FALSE))
   }
 
-  invisible()
+  args <- c(extension, "--no-prompt", if (quiet) cli_arg_quiet(), quarto_args)
+  quarto_update(args, quarto_bin = quarto_bin, echo = TRUE)
+  if (!quiet) {
+    cli::cli_inform("Extension {.code {extension}} updated.")
+  }
+  invisible(TRUE)
 }
 
 quarto_update <- function(args = character(), ...) {